backend/core_saas/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
Django==4.2.27
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg==3.2.13
psycopg-binary==3.2.13
PyJWT==2.9.0
python-dotenv==1.0.1
redis==6.1.1
sqlparse==0.5.4
typing_extensions==4.13.2
